+++ photutils/psf/epsf_builder.py
@@ -5,6 +5,7 @@ King 2000 (PASP 112, 1360) and Anderson 2016 (WFC3 ISR 2016-12).
 """
 
 import copy
+import inspect
 import warnings
 from dataclasses import dataclass
 
@@ -12,7 +13,9 @@ import numpy as np
 from astropy.modeling.fitting import TRFLSQFitter
 from astropy.nddata import NoOverlapError, PartialOverlapError, overlap_slices
 from astropy.stats import SigmaClip
-from astropy.utils.exceptions import AstropyUserWarning
+from astropy.utils.decorators import deprecated
+from astropy.utils.exceptions import (AstropyDeprecationWarning,
+                                      AstropyUserWarning)
 from scipy.ndimage import convolve
 
 from photutils.centroids import centroid_com
@@ -715,9 +718,18 @@ class EPSFBuildResult:
         raise IndexError(msg)
 
 
+@deprecated(since='3.0.0',
+            message=('EPSFFitter is deprecated and will be removed in a '
+                     'future version. Use EPSFBuilder with the fitter, '
+                     'fit_shape, and fitter_maxiters parameters instead.'))
 class EPSFFitter:
     """
     Class to fit an ePSF model to one or more stars.
+
+    .. deprecated:: 3.0.0
+        `EPSFFitter` is deprecated. Use `EPSFBuilder` with the
+        ``fitter``, ``fit_shape``, and ``fitter_maxiters`` parameters
+        instead.
 
     Parameters
     ----------
@@ -971,11 +983,32 @@ class EPSFBuilder:
         stars must meet this condition for the building iterations to
         stop.
 
-    fitter : `EPSFFitter` object, optional
-        A `EPSFFitter` object use to fit the ePSF to stars. If `None`,
-        then the default `EPSFFitter` will be used. To set custom fitter
-        options, input a new `EPSFFitter` object. See the `EPSFFitter`
-        documentation for options.
+    fitter : `~astropy.modeling.fitting.Fitter` or `EPSFFitter`, optional
+        A `~astropy.modeling.fitting.Fitter` object used to fit the
+        ePSF to stars. If `None`, then the default
+        `~astropy.modeling.fitting.TRFLSQFitter` will be used.
+
+        .. deprecated:: 3.0.0
+            Passing an `EPSFFitter` instance is deprecated; use
+            the ``fitter``, ``fit_shape``, and ``fitter_maxiters``
+            parameters instead.
+
+    fit_shape : int, tuple of int, or `None`, optional
+        The size (in pixels) of the box centered on the star to be
+        used for ePSF fitting. This allows using only a small number
+        of central pixels of the star (i.e., where the star is
+        brightest) for fitting. If ``fit_shape`` is a scalar then a
+        square box of size ``fit_shape`` will be used. If ``fit_shape``
+        has two elements, they must be in ``(ny, nx)`` order.
+        ``fit_shape`` must have odd values and be greater than or equal
+        to 3 for both axes. If `None`, the fitter will use the entire
+        star image.
+
+    fitter_maxiters : int, optional
+        The maximum number of iterations in which the ``fitter`` is
+        called for each star. The value can be increased if the fit
+        is not converging. This parameter is passed to the ``fitter``
+        if it supports the ``maxiter`` parameter and ignored otherwise.
 
     maxiters : int, optional
         The maximum number of ePSF building iterations to perform.
@@ -997,7 +1030,8 @@ class EPSFBuilder:
                  smoothing_kernel='quartic', sigma_clip=SIGMA_CLIP,
                  recentering_func=centroid_com, recentering_boxsize=(5, 5),
                  recentering_maxiters=20, center_accuracy=1.0e-3,
-                 fitter=None, maxiters=10, progress_bar=True):
+                 fitter=None, fit_shape=5, fitter_maxiters=100,
+                 maxiters=10, progress_bar=True):
 
         # Validate and store oversampling using the validator
         self.oversampling = _EPSFValidator.validate_oversampling(
@@ -1018,12 +1052,44 @@ class EPSFBuilder:
                                            lower_bound=(3, 0), check_odd=True)
         self.smoothing_kernel = smoothing_kernel
 
-        if fitter is None:
-            fitter = EPSFFitter()
-        if not isinstance(fitter, EPSFFitter):
-            msg = 'fitter must be an EPSFFitter instance'
-            raise TypeError(msg)
-        self.fitter = fitter
+        # Handle fitter parameter - accept both astropy Fitter and
+        # deprecated EPSFFitter for backward compatibility
+        if isinstance(fitter, EPSFFitter):
+            warnings.warn('Passing an EPSFFitter instance to '
+                          'EPSFBuilder is deprecated. Use the fitter, '
+                          'fit_shape, and fitter_maxiters parameters '
+                          'instead.',
+                          AstropyDeprecationWarning)
+            self.fitter = fitter.fitter
+            self.fit_shape = fitter.fit_boxsize
+            self.fitter_maxiters = None
+            self._fitter_kwargs = fitter.fitter_kwargs
+        else:
+            if fitter is None:
+                fitter = TRFLSQFitter()
+            if not callable(fitter):
+                msg = 'fitter must be a callable astropy Fitter instance'
+                raise TypeError(msg)
+            self.fitter = fitter
+
+            # Validate fit_shape
+            if fit_shape is not None:
+                self.fit_shape = as_pair('fit_shape', fit_shape,
+                                         lower_bound=(3, 0),
+                                         check_odd=True)
+            else:
+                self.fit_shape = None
+
+            # Validate fitter_maxiters
+            self.fitter_maxiters = self._validate_fitter_maxiters(
+                fitter_maxiters)
+
+            # Build fitter keyword arguments
+            self._fitter_kwargs = {}
+            if self.fitter_maxiters is not None:
+                self._fitter_kwargs['maxiter'] = self.fitter_maxiters
+
+        self._fitter_has_fit_info = hasattr(self.fitter, 'fit_info')
 
         # Validate center accuracy using the validator
         _EPSFValidator.validate_center_accuracy(center_accuracy)
@@ -1061,6 +1127,32 @@ class EPSFBuilder:
             The result of the ePSF building process.
         """
         return self._build_epsf(stars)
+
+    def _validate_fitter_maxiters(self, fitter_maxiters):
+        """
+        Validate the ``fitter_maxiters`` parameter.
+
+        Parameters
+        ----------
+        fitter_maxiters : int
+            Maximum number of fitter iterations to validate.
+
+        Returns
+        -------
+        fitter_maxiters : int or `None`
+            The validated value, or `None` if the fitter does not
+            support the ``maxiter`` parameter.
+        """
+        spec = inspect.signature(self.fitter.__call__)
+        has_maxiter = ('maxiter' in spec.parameters
+                       or any(p.kind == inspect.Parameter.VAR_KEYWORD
+                              for p in spec.parameters.values()))
+        if not has_maxiter:
+            warnings.warn('"fitter_maxiters" will be ignored because '
+                          'it is not accepted by the input fitter',
+                          AstropyUserWarning)
+            return None
+        return fitter_maxiters
 
     def _create_initial_epsf(self, stars):
         """
@@ -1571,6 +1663,155 @@ class EPSFBuilder:
 
         return converged, center_dist_sq, new_centers
 
+    def _fit_stars(self, epsf, stars):
+        """
+        Fit an ePSF model to stars.
+
+        Parameters
+        ----------
+        epsf : `ImagePSF`
+            An ePSF model to be fitted to the stars.
+
+        stars : `EPSFStars` object
+            The stars to be fit. The center coordinates for each star
+            should be as close as possible to actual centers. For stars
+            that contain weights, a weighted fit of the ePSF to the
+            star will be performed.
+
+        Returns
+        -------
+        fitted_stars : `EPSFStars` object
+            The fitted stars. The ePSF-fitted center position and flux
+            are stored in the ``center`` (and ``cutout_center``) and
+            ``flux`` attributes.
+        """
+        if len(stars) == 0:
+            return stars
+
+        if not isinstance(epsf, ImagePSF):
+            msg = 'The input epsf must be an ImagePSF'
+            raise TypeError(msg)
+
+        fitted_stars = []
+        for star in stars:
+            if isinstance(star, EPSFStar):
+                if star._excluded_from_fit:
+                    fitted_star = star
+                else:
+                    fitted_star = self._fit_star(epsf, star)
+
+            elif isinstance(star, LinkedEPSFStar):
+                fitted_star = []
+                for linked_star in star:
+                    if linked_star._excluded_from_fit:
+                        fitted_star.append(linked_star)
+                    else:
+                        fitted_star.append(self._fit_star(epsf,
+                                                          linked_star))
+
+                fitted_star = LinkedEPSFStar(fitted_star)
+                fitted_star.constrain_centers()
+
+            else:
+                msg = ('stars must contain only EPSFStar and/or '
+                       'LinkedEPSFStar objects')
+                raise TypeError(msg)
+
+            fitted_stars.append(fitted_star)
+
+        return EPSFStars(fitted_stars)
+
+    def _fit_star(self, epsf, star):
+        """
+        Fit an ePSF model to a single star.
+
+        Parameters
+        ----------
+        epsf : `ImagePSF`
+            An ePSF model to be fitted to the star.
+
+        star : `EPSFStar`
+            The star to be fit.
+
+        Returns
+        -------
+        star : `EPSFStar`
+            The fitted star with updated cutout center and flux.
+        """
+        fit_shape = self.fit_shape
+        fitter = self.fitter
+        fitter_kwargs = self._fitter_kwargs
+        fitter_has_fit_info = self._fitter_has_fit_info
+
+        if fit_shape is not None:
+            try:
+                xcenter, ycenter = star.cutout_center
+                large_slc, _ = overlap_slices(star.shape, fit_shape,
+                                              (ycenter, xcenter),
+                                              mode='strict')
+            except (PartialOverlapError, NoOverlapError):
+                star = copy.copy(star)
+                star._fit_error_status = 1
+                return star
+
+            data = star.data[large_slc]
+            weights = star.weights[large_slc]
+
+            # Define the origin of the fitting region
+            x0 = large_slc[1].start
+            y0 = large_slc[0].start
+        else:
+            # Use the entire cutout image
+            data = star.data
+            weights = star.weights
+
+            # Define the origin of the fitting region
+            x0 = 0
+            y0 = 0
+
+        # Define positions in the undersampled grid. The fitter will
+        # evaluate on the defined interpolation grid, currently in the
+        # range [0, len(undersampled grid)].
+        yy, xx = np.indices(data.shape, dtype=float)
+        xx = xx + x0 - star.cutout_center[0]
+        yy = yy + y0 - star.cutout_center[1]
+
+        # Define the initial guesses for fitted flux and shifts
+        epsf.flux = star.flux
+        epsf.x_0 = 0.0
+        epsf.y_0 = 0.0
+
+        try:
+            fitted_epsf = fitter(model=epsf, x=xx, y=yy, z=data,
+                                 weights=weights, **fitter_kwargs)
+        except TypeError:
+            # Handle case where the fitter does not support weights
+            fitted_epsf = fitter(model=epsf, x=xx, y=yy, z=data,
+                                 **fitter_kwargs)
+
+        fit_error_status = 0
+        if fitter_has_fit_info:
+            fit_info = copy.copy(fitter.fit_info)
+            if 'ierr' in fit_info and fit_info['ierr'] not in [1, 2, 3, 4]:
+                fit_error_status = 2  # fit solution was not found
+        else:
+            fit_info = None
+
+        # Compute the star's fitted position
+        x_center = star.cutout_center[0] + fitted_epsf.x_0.value
+        y_center = star.cutout_center[1] + fitted_epsf.y_0.value
+
+        star = copy.copy(star)
+        star.cutout_center = (x_center, y_center)
+
+        # Set the star's flux to the ePSF-fitted flux
+        star.flux = fitted_epsf.flux.value
+
+        star._fit_info = fit_info
+        star._fit_error_status = fit_error_status
+
+        return star
+
     def _process_iteration(self, stars, epsf, iter_num):
         """
         Process a single iteration of ePSF building.
@@ -1606,7 +1847,7 @@ class EPSFBuilder:
             warnings.filterwarnings('ignore', message=message,
                                     category=AstropyUserWarning)
 
-            stars = self.fitter(epsf, stars)
+            stars = self._fit_stars(epsf, stars)
 
         # Reset ePSF flux to 1.0 after fitting (fitting modifies the
         # flux)